Adapt to 2.12.1
Use 2.11 interface for 2.12.0, as it's identical.

From 2.12.1 on, there's a hook method that can be
used instead of copy/paste polymorphism.

One test failure due to a change in how members are sorted.

PS: since sbt 0.13.8, there's no need to manually add
versioned source files. Sadly, we need a special case for 2.12.0