handle more Scala version numbers correctly
annoyingly, a Scala built by the PR validator has an artifact
version of e.g. `2.13.0-pre-c6065cc-SNAPSHOT` but a
versionNumberString of e.g. 2.13.0-20180218-172225-c6065cc`.
the old logic was handling this incorrectly.

with this change, we lose the ability to distinguish between
e.g. 2.13.0-M2 and 2.13.0-M3.  but it's fine, it's rare we actually
need to change the patch file, and when a new Scala milestone comes
out, it's fine to drop support for the previous milestone