Don't run genjavadoc while generating scaladoc
When generating scaladoc, not the scala compiler is run with a limited set of
phases. Notably, it runs it without the 'fields' phase which genjavadoc
depends on, disabling genjavadoc while producing the warning
`dropping dependency on node with no phase object: fields`. When you have
`-Xfatal-warnings` enabled, this then leads to an error.

Inspired by https://github.com/HairyFotr/linter/issues/23