Set compiler phase `runsBefore` to `tailcalls`
`runsAfter` is set to `fields`. However, in reality the phase
assembly algorithm used in 2.12.8 and before, also 2.13.0, places the
`GenJavadoc` phase right after `specialize`, which is valid, but probably
not intended.

With some changes in the phase assembly algorithm in 2.13.1, the
`GenJavadoc` phase was now placed right after `fields`, which caused
tests to fail.