Update to sbt 1.3.3 and use sbt-ci-release (#208)
References #193, #203.
The issue with publishLocal mentioned in #193 seems to have disappeared.

This should also use sonatype bundle uploads for releasing, which might
speed up the travis release enough so it completes within the deadline.
Still needs to be tested, though, because that feature is only used when
doing non-snapshot releases. We should probably release an RC first to
test the process.